Refactor App and Menu components into separate files
- Extract Header and Menu components from App.jsx
- Remove redundant state management and scroll handling logic
- Simplify App.jsx by moving complex UI elements to dedicated components
- Clean up unused CSS styles in App.css
