+++ src/App.jsx
@@ -1,16 +1,12 @@
 import { useState, useEffect } from 'react'
 import { Routes, Route, useNavigate, useLocation } from 'react-router-dom'
 import './App.css'
-import logo from './assets/images/logo.png'
-import { AiOutlineHome } from 'react-icons/ai'
+import Menu from './components/Menu'
 import { 
-  MdOutlineDesignServices,
-  MdDevices,
   MdSpeed,
   MdSecurity,
   MdOutlineSupport
 } from 'react-icons/md'
-import { BsPeople } from 'react-icons/bs'
 import { 
   FaAndroid,
   FaApple,
@@ -21,12 +17,11 @@ import AboutUs from './Pages/AboutUs'
 import Services from './Pages/Services'
 import Footer from './components/Footer'
 import ServicesInfo from './Pages/ServicesInfo'
+import Header from './components/Header'
 
 function App() {
   const [isMenuOpen, setIsMenuOpen] = useState(false)
   const [activeSection, setActiveSection] = useState(0)
-  const [isScrolled, setIsScrolled] = useState(false)
-  const [scrollPosition, setScrollPosition] = useState(0)
   const [currentLanguage, setCurrentLanguage] = useState('sv')
   const navigate = useNavigate()
   const location = useLocation()
@@ -47,55 +42,17 @@ function App() {
   }, [])
 
   useEffect(() => {
-    if (isMenuOpen) {
-      document.body.classList.add('body-locked')
-    } else {
-      document.body.classList.remove('body-locked')
-    }
-    return () => document.body.classList.remove('body-locked')
-  }, [isMenuOpen])
-
-  useEffect(() => {
-    const handleScroll = () => {
-      if (window.scrollY > 20) {
-        setIsScrolled(true)
-      } else {
-        setIsScrolled(false)
-      }
-    }
-
-    window.addEventListener('scroll', handleScroll)
-    return () => window.removeEventListener('scroll', handleScroll)
-  }, [])
-
-  useEffect(() => {
-    if (isMenuOpen) {
-      document.documentElement.style.setProperty('--scroll-position', `${scrollPosition}px`);
-    } else {
-      document.documentElement.style.removeProperty('--scroll-position');
-      window.scrollTo(0, scrollPosition);
-    }
-  }, [isMenuOpen, scrollPosition]);
-
-  useEffect(() => {
     window.scrollTo(0, 0)
   }, [location.pathname])
 
   const toggleMenu = () => {
-    if (!isMenuOpen) {
-      setScrollPosition(window.scrollY)
-    }
     setIsMenuOpen(!isMenuOpen)
-  }
-  const handleNavigation = (path) => {
-    navigate(path)
-    setIsMenuOpen(false)
   }
 
   const handleLanguageChange = (lang) => {
-    setCurrentLanguage(lang);
-    toggleMenu();
-  };
+    setCurrentLanguage(lang)
+    toggleMenu()
+  }
 
   const HomePage = () => (
     <div className="home-content">
@@ -273,56 +230,13 @@ function App() {
 
   return (
     <div className={`app-wrapper ${isMenuOpen ? 'menu-open' : ''}`}>
-      <div className={`overlay ${isMenuOpen ? 'open' : ''}`} onClick={toggleMenu}></div>
-      <header className={`header ${isScrolled ? 'scrolled' : ''}`}>
-        <img src={logo} alt="Logo" className="logo" onClick={() => handleNavigation('/')} />
-        <div className={`menu-icon ${isMenuOpen ? 'open' : ''}`} onClick={toggleMenu}>
-          <div className="menu-line"></div>
-          <div className="menu-line"></div>
-          <div className="menu-line"></div>
-        </div>
-      </header>
-      <nav className={`menu ${isMenuOpen ? 'open' : ''}`}>
-        <button className="menu-close-btn" onClick={toggleMenu}>×</button>
-        <ul className="menu-items">
-          <li onClick={() => handleNavigation('/')}>
-            <AiOutlineHome className="menu-icon-item" />
-            <span>Hem</span>
-          </li>
-          <li onClick={() => handleNavigation('/services')}>
-            <MdOutlineDesignServices className="menu-icon-item" />
-            <span>Tjänster</span>
-          </li>
-          <li onClick={() => handleNavigation('/prices')}>
-            <FaMobileAlt className="menu-icon-item" />
-            <span>Priser</span>
-          </li>
-          <li onClick={() => handleNavigation('/about')}>
-            <BsPeople className="menu-icon-item" />
-            <span>Om oss</span>
-          </li>
-          <li className="language-selector">
-            <div className="language-container">
-              <div className="language-label">Språk:</div>
-              <div className="language-options">
-                <span 
-                  className={`language-option ${currentLanguage === 'sv' ? 'active' : ''}`}
-                  onClick={() => handleLanguageChange('sv')}
-                >
-                  Svenska
-                </span>
-                <span className="separator">|</span>
-                <span 
-                  className={`language-option ${currentLanguage === 'en' ? 'active' : ''}`}
-                  onClick={() => handleLanguageChange('en')}
-                >
-                  English
-                </span>
-              </div>
-            </div>
-          </li>
-        </ul>
-      </nav>
+      <Header isMenuOpen={isMenuOpen} toggleMenu={toggleMenu} />
+      <Menu 
+        isMenuOpen={isMenuOpen}
+        toggleMenu={toggleMenu}
+        currentLanguage={currentLanguage}
+        handleLanguageChange={handleLanguageChange}
+      />
       <div className="card">
         <Routes>
           <Route path="/" element={<HomePage />} />
