+++ src/components/Header.css
@@ -1,0 +1,73 @@
+.header {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  height: 4rem;
+  z-index: 1002;
+  transition: all 0.3s ease;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.header.scrolled {
+  background: rgba(255, 255, 255, 0.95);
+  backdrop-filter: blur(10px);
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+}
+
+.logo {
+  position: fixed;
+  top: 0.75rem;
+  left: 1rem;
+  height: 45px;
+  width: auto;
+  object-fit: contain;
+  z-index: inherit;
+  cursor: pointer;
+  transition: transform 0.3s ease;
+}
+
+.logo:hover {
+  transform: scale(1.05);
+}
+
+/* Media Queries */
+@media (min-width: 600px) {
+  .logo {
+    height: 48px;
+  }
+}
+
+@media (min-width: 768px) {
+  .logo {
+    height: 50px;
+  }
+  .menu-line {
+    width: 30px;
+  }
+}
+
+@media (min-width: 1280px) {
+  .logo {
+    top: 0.5rem;
+    height: 55px;
+  }
+  .menu-line {
+    width: 35px;
+  }
+}
+
+@media (min-width: 1600px) {
+  .logo {
+    top: 0.4rem;
+    height: 58px;
+  }
+}
+
+@media (min-width: 2560px) {
+  .logo {
+    height: 60px;
+  }
+}
