+++ src/components/Menu.css
@@ -1,0 +1,254 @@
+/* Copy these styles from App.css */
+.menu {
+  position: fixed;
+  top: 0;
+  right: -100%;
+  width: 40%;
+  height: 100vh;
+  background: rgba(255, 255, 255, 0.95);
+  padding: 6rem 2rem 2rem 2rem;
+  box-shadow: -4px 0 6px rgba(0, 0, 0, 0.1);
+  transition: right 0.3s ease-in-out;
+  z-index: 95;
+  overflow-y: auto;
+}
+
+.menu.open {
+  right: 0;
+  z-index: 999;
+}
+
+.overlay {
+  display: none;
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.5);
+  z-index: 90;
+}
+
+.overlay.open {
+  display: block;
+}
+
+/* For tablets and larger screens */
+@media (min-width: 768px) {
+  .menu {
+    width: 25%;
+  }
+}
+
+/* For desktop screens */
+@media (min-width: 1280px) {
+  .menu {
+    width: 20%;
+  }
+}
+
+/* For very large screens */
+@media (min-width: 1600px) {
+  .menu {
+    width: 15%;
+  }
+}
+
+.menu-items {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 0;
+}
+
+.menu-items li {
+  padding: 0.8rem 1.2rem;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  font-size: 0.95rem;
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+  border-left: 3px solid transparent;
+  background: transparent;
+  box-shadow: none;
+  border-bottom: 1px solid rgba(0, 0, 0, 0.1);
+}
+
+.menu-items li:last-child {
+  border-bottom: none;
+}
+
+.menu-items li:hover {
+  background: rgba(100, 108, 255, 0.1);
+  border-left: 3px solid #646cff;
+  transform: none;
+  box-shadow: none;
+}
+
+.menu-icon-item {
+  font-size: 1.3rem;
+  color: #646cff;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.menu-items li span {
+  font-weight: 500;
+  color: #213547;
+  letter-spacing: 0.3px;
+}
+
+.menu-icon {
+  position: fixed;
+  top: 1.5rem;
+  right: 1.1rem;
+  cursor: pointer;
+  z-index: 1003;
+  width: 30px;
+  height: 22px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.menu-line {
+  width: 100%;
+  height: 3px;
+  background-color: #213547;
+  margin: 0;
+  transition: all 0.3s ease;
+}
+
+/* Menu icon animation when open */
+.menu-icon.open .menu-line:nth-child(1) {
+  transform: rotate(-45deg) translate(-6px, 7px);
+  background-color: #213547;
+}
+
+.menu-icon.open .menu-line:nth-child(2) {
+  opacity: 0;
+  transform: translateX(-20px);
+}
+
+.menu-icon.open .menu-line:nth-child(3) {
+  transform: rotate(45deg) translate(-6px, -7px);
+  background-color: #213547;
+}
+
+/* Adjust sizes for larger screens */
+@media (min-width: 768px) {
+  .menu-items li {
+    padding: 0.9rem 1.5rem;
+    font-size: 1rem;
+  }
+  
+  .menu-icon-item {
+    font-size: 1.4rem;
+  }
+}
+
+@media (min-width: 1280px) {
+  .menu-items li {
+    font-size: 1.1rem;
+  }
+  
+  .menu-icon-item {
+    font-size: 1.5rem;
+  }
+}
+
+/* Mobile adjustments */
+@media (max-width: 768px) {
+  .menu {
+    width: 75%;
+  }
+  
+  .menu-icon {
+    margin-right: 0.5rem;
+  }
+}
+
+/* Add language selector styles */
+.language-selector {
+  flex-direction: row !important;
+  align-items: center !important;
+  justify-content: center !important;
+  border-left: none !important;
+  margin-top: 1rem;
+}
+
+.language-selector:hover {
+  background: transparent !important;
+  border-left: none !important;
+  transform: none !important;
+  box-shadow: none !important;
+}
+
+.language-container {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+  justify-content: center;
+  width: 100%;
+}
+
+.language-label {
+  color: #213547;
+  font-weight: 500;
+  white-space: nowrap;
+}
+
+.language-options {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.language-option {
+  color: #213547;
+  cursor: pointer;
+  padding-bottom: 0.2rem;
+  transition: all 0.3s ease;
+}
+
+.language-option:hover {
+  color: #646cff;
+}
+
+.language-option.active {
+  border-bottom: 2px solid #646cff;
+  color: #646cff;
+}
+
+.language-options .separator {
+  color: #666;
+}
+
+.menu-close-btn {
+  position: absolute;
+  top: 1rem;
+  right: 1rem;
+  background: none;
+  border: none;
+  font-size: 2.5rem;
+  font-weight: 300;
+  cursor: pointer;
+  color: #646cff;
+  padding: 0.5rem;
+  z-index: 1002;
+  transition: transform 0.3s ease;
+  width: 50px;
+  height: 50px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.menu-close-btn:hover {
+  transform: scale(1.1);
+  color: #535bf2;
+}
